+++ pai-dlc-20201203/alibabacloud_pai_dlc20201203/models/_autoscaling_metric_spec.py
@@ -8,9 +8,11 @@ class AutoscalingMetricSpec(DaraModel):
     def __init__(
         self,
         metric_name: str = None,
+        stabilization_window_seconds: int = None,
         target_value: int = None,
     ):
         self.metric_name = metric_name
+        self.stabilization_window_seconds = stabilization_window_seconds
         self.target_value = target_value
 
     def validate(self):
@@ -24,6 +26,9 @@ class AutoscalingMetricSpec(DaraModel):
         if self.metric_name is not None:
             result['MetricName'] = self.metric_name
 
+        if self.stabilization_window_seconds is not None:
+            result['StabilizationWindowSeconds'] = self.stabilization_window_seconds
+
         if self.target_value is not None:
             result['TargetValue'] = self.target_value
 
@@ -33,6 +38,9 @@ class AutoscalingMetricSpec(DaraModel):
         m = m or dict()
         if m.get('MetricName') is not None:
             self.metric_name = m.get('MetricName')
+
+        if m.get('StabilizationWindowSeconds') is not None:
+            self.stabilization_window_seconds = m.get('StabilizationWindowSeconds')
 
         if m.get('TargetValue') is not None:
             self.target_value = m.get('TargetValue')