+++ pai-dlc-20201203/alibabacloud_pai_dlc20201203/models/_get_job_response_body.py
@@ -44,6 +44,7 @@ class GetJobResponseBody(DaraModel):
         restart_record: List[main_models.GetJobResponseBodyRestartRecord] = None,
         restart_times: str = None,
         role_system_envs: Dict[str, dict] = None,
+        scheduling_strategy: str = None,
         settings: main_models.JobSettings = None,
         status: str = None,
         status_history: List[main_models.StatusTransitionItem] = None,
@@ -123,6 +124,7 @@ class GetJobResponseBody(DaraModel):
         # The number of retries and the maximum number of retries used by the job.
         self.restart_times = restart_times
         self.role_system_envs = role_system_envs
+        self.scheduling_strategy = scheduling_strategy
         # The additional parameter configurations of the job.
         self.settings = settings
         # The status of the job. Valid values:
@@ -321,6 +323,9 @@ class GetJobResponseBody(DaraModel):
         if self.role_system_envs is not None:
             result['RoleSystemEnvs'] = self.role_system_envs
 
+        if self.scheduling_strategy is not None:
+            result['SchedulingStrategy'] = self.scheduling_strategy
+
         if self.settings is not None:
             result['Settings'] = self.settings.to_map()
 
@@ -485,6 +490,9 @@ class GetJobResponseBody(DaraModel):
 
         if m.get('RoleSystemEnvs') is not None:
             self.role_system_envs = m.get('RoleSystemEnvs')
+
+        if m.get('SchedulingStrategy') is not None:
+            self.scheduling_strategy = m.get('SchedulingStrategy')
 
         if m.get('Settings') is not None:
             temp_model = main_models.JobSettings()